feat(genai): Add generate content with routing option

+++ genai/tools/tools_code_exec_with_txt.go
@@ -45,7 +45,7 @@ func generateWithCodeExec(w io.Writer) error {
 		Tools: []*genai.Tool{
 			{CodeExecution: &genai.ToolCodeExecution{}},
 		},
-		Temperature: genai.Ptr(0.0),
+		Temperature: genai.Ptr(float32(0.0)),
 	}
 	modelName := "gemini-2.0-flash-001"
